+++ jib-gradle-plugin/src/test/java/com/google/cloud/tools/jib/gradle/JibExtensionTest.java
@@ -96,6 +96,14 @@ public class JibExtensionTest {
   }
 
   @Test
+  public void testArgs() {
+    Assert.assertEquals(Collections.emptyList(), testJibExtension.getArgs());
+
+    testJibExtension.setArgs(Arrays.asList("arg1", "arg2"));
+    Assert.assertEquals(Arrays.asList("arg1", "arg2"), testJibExtension.getArgs());
+  }
+
+  @Test
   public void testFormat() {
     Assert.assertEquals(V22ManifestTemplate.class, testJibExtension.getFormat());